Update FmcLinkController port voltage algorithm
- Expose more parameters of its operation
- Final link check after setting offset voltage
- From testing it was found that headstage-64 needs a very large
amount of time to reach an acceptable votlage level to be considered off
(nearly 10 seconds)
- A new virtual SetPortVoltage method was introduced in the base link
controller to allow headstages to change the procedure depending on its
needs. In the case of headstage-64, a two-step procedure with a long
wait time after turning port voltage off, and a very large offset
voltage applied after lock.