Add support for UCLA Miniscope V4
- This device needs some work arounds because its clock is completely
out of spec for the SERDES it uses. The back channel must be used to
configure a PLL on the camera in order to bump the PCLK rate before it
will like. For this reason, I mad ConfigureFmcLinkController.
CheckLinkState virtual so I could override its behavior with these steps
before checking the link state
- Added configuraiton and streaming functionality.
- There are few outstanding issues:
1. Something is wrong with the DS90UB9x configuration resulting in 1
missing column from each frame and imaging data that "slides" by one
column for each image produced
2. The axis map for the BNO needs to be set correctly
3. I was unable to verify that the EWL was functioning. It might be
fine, but I was unable to verify.